# prompt/bright/trustivs_openapi_3_1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrustIVS AIoT 管理平台 OpenAPI",
    "summary": "基于《TrustIVS AIoT 管理平台 OpenAPI V2.0》整理的 OpenAPI 3.1 规范。",
    "version": "2.0-derived",
    "description": "本规范由 PDF 文档（TrustIVS AIoT 管理平台 OpenAPI V2.0）梳理而成，覆盖鉴权、AI 抓拍、回放与直播、在线设备/通道、设备/通道配置、人脸库/人脸信息、人脸识别记录、车辆库/车辆信息、车牌识别记录、心跳上报、环境监控上报、监管单位、机构等模块。",
    "contact": {
      "name": "深圳市亚略特科技股份有限公司",
      "url": "https://www.aratek.com.cn"
    },
    "x-generated-at": "2025-09-04T02:28:38.743660"
  },
  "servers": [
    {
      "url": "http://{ip}:{port}",
      "description": "HTTP 服务器",
      "variables": {
        "ip": {
          "default": "127.0.0.1"
        },
        "port": {
          "default": "9086"
        }
      }
    },
    {
      "url": "https://{ip}:{port}",
      "description": "HTTPS 服务器",
      "variables": {
        "ip": {
          "default": "127.0.0.1"
        },
        "port": {
          "default": "443"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "会话 Token"
      }
    },
    "parameters": {
      "HdrTime": {
        "name": "time",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "时间戳"
      },
      "HdrUUID": {
        "name": "uuid",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "UUID"
      }
    },
    "schemas": {
      "StdResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "1 成功，0 失败"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        },
        "required": [
          "code"
        ]
      },
      "Paged": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "list": {
            "type": "array",
            "items": {}
          }
        }
      },
...
      "Company": {
        "type": "object",
        "properties": {
          "fnumber": {
            "type": "string"
          },
          "fname": {
            "type": "string"
          },
          "fcity": {
            "type": "string"
          },
          "fprovince": {
            "type": "string"
          },
          "fregion": {
            "type": "string"
          },
          "fmobile": {
            "type": "string"
          },
          "fcontacts": {
            "type": "string"
          },
          "faddress": {
            "type": "string"
          },
          "fsimplename": {
            "type": "string"
          },
          "fsocialcreditcode": {
            "type": "string"
          },
          "fregulatorname": {
            "type": "string"
          },
          "flongitude": {
            "type": "number"
          },
          "fdimension": {
            "type": "number"
          },
          "fenttype": {
            "type": "integer",
            "description": "0餐饮 1学校 2加油站 3冷库 4社区"
          },
          "fplatacc": {
            "type": "string"
          },
          "fthirdcomnum": {
            "type": "string"
          }
        }
      },
      "Camera": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "fname": {
            "type": "string"
          },
          "deviceHostIp": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ON",
              "OFF",
              "on",
              "off"
            ]
          },
          "deviceSn": {
            "type": "string"
          },
          "updateTime": {
            "type": "string"
          },
          "fnationcode": {
            "type": "string"
          },
          "flv": {
            "type": "string"
          },
          "hls": {
            "type": "string"
          },
          "rtmp": {
            "type": "string"
          },
          "webrtc": {
            "type": "string"
          },
          "ws_flv": {
            "type": "string"
          }
        }
      },
      "AIRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fsocialcreditcode": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "thumbnailBase64": {
            "type": "string"
          },
          "originalBase64": {
            "type": "string"
          },
          "fimgpath": {
            "type": "string"
          },
          "foriginimgpath": {
            "type": "string"
          },
          "createTime": {
            "type": "string"
          },
          "nvrsn": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "equipId": {
            "type": "string"
          },
          "mp4Path": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "fnumber": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiToken": []
    }
  ],
  "paths": {
...
    "/gatewayGBS/openApi/getCompanyList": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "获取机构信息",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fplatacc": {
                    "type": "string",
                    "description": "第三方平台账号"
                  },
                  "regulatorSocialCreditCode": {
                    "type": "string",
                    "description": "监管单位社会统一信用代码"
                  },
                  "pageSize": {
                    "type": "integer",
                    "default": 10,
                    "maximum": 50
                  },
                  "pageNum": {
                    "type": "integer",
                    "default": 1
                  }
                },
                "required": [
                  "fplatacc"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StdResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Paged"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "list": {
                                  "items": {
                                    "$ref": "#/components/schemas/Company"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getCameraByCompany": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "根据机构查询摄像头与直播地址",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fthirdcomnum": {
                    "type": "string",
                    "description": "第三方机构编码"
                  },
                  "deviceSn": {
                    "type": "string",
                    "description": "设备 SN"
                  },
                  "pageSize": {
                    "type": "integer",
                    "maximum": 100,
                    "default": 10
                  },
                  "pageNum": {
                    "type": "integer",
                    "default": 1
                  }
                },
                "required": [
                  "fthirdcomnum",
                  "deviceSn"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StdResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Paged"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "list": {
                                  "items": {
                                    "$ref": "#/components/schemas/Camera"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getAIRecordByCompany": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "根据机构查询 AI 抓拍记录",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fthirdcomnum": {
                    "type": "string"
                  },
                  "fsocialcreditcode": {
                    "type": "string"
                  },
                  "nvrsn": {
                    "type": "string"
                  },
                  "typeId": {
                    "type": "integer"
                  },
                  "page": {
                    "type": "integer",
                    "default": 1
                  },
                  "pageSize": {
                    "type": "integer",
                    "default": 10
                  },
                  "timeStart": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timeEnd": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "anyOf": [
                  {
                    "required": [
                      "fthirdcomnum"
                    ]
                  },
                  {
                    "required": [
                      "fsocialcreditcode"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StdResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Paged"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "list": {
                                  "items": {
                                    "$ref": "#/components/schemas/AIRecord"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/openApi/sendAIRecordByCompany": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "AI 抓拍信息推送（第三方接收端）",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AIRecord"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getBackUrl": {
      "get": {
        "tags": [
          "AI Capture"
        ],
        "summary": "获取回放地址",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          },
          {
            "name": "serial",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "starttime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endtime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StdResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "rtmp": {
                              "type": "string"
                            },
                            "flv": {
                              "type": "string"
                            },
                            "webrtc": {
                              "type": "string"
                            },
                            "ws_flv": {
                              "type": "string"
                            },
                            "hls": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/gatewayGBS/openApi/getDownloadUrl": {
      "get": {
        "tags": [
          "AI Capture"
        ],
        "summary": "获取回放下载地址",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          },
          {
            "name": "serial",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "starttime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endtime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getStreamURL": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "获取通道直播地址",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "unitSocailCreditCode": {
                    "type": "string"
                  },
                  "deviceId": {
                    "type": "string"
                  },
                  "channelId": {
                    "type": "string"
                  }
                },
                "required": [
                  "unitSocailCreditCode",
                  "deviceId",
                  "channelId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getSnap": {
      "get": {
        "tags": [
          "AI Capture"
        ],
        "summary": "获取通道快照",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          },
          {
            "name": "deviceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getDeviceListByCompany": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "根据机构查询设备列表",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fsocialcreditcode": {
                    "type": "string"
                  }
                },
                "required": [
                  "fsocialcreditcode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gatewayGBS/openApi/getChannelByDevice": {
      "post": {
        "tags": [
          "AI Capture"
        ],
        "summary": "根据设备国标号获取通道列表",
        "parameters": [
          {
            "$ref": "#/components/parameters/HdrTime"
          },
          {
            "$ref": "#/components/parameters/HdrUUID"
          }
        ],
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deviceId": {
                    "type": "string"
                  }
                },
                "required": [
                  "deviceId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StdResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}